requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.3
CacheControl==0.14.4
cachetools==6.2.2
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==46.0.3
Django==5.2.8
django-allauth==65.13.0
django-cors-headers==4.6.0
django-debug-toolbar==6.1.0
django-extensions==4.1
django-filter==25.1
django-redis==6.0.0
django-request-id==1.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.11
fcm-django==2.3.1
firebase_admin==7.1.0
frozenlist==1.8.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.2
MarkupSafe==3.0.3
more-itertools==10.8.0
msgpack==1.1.2
multidict==6.7.0
packaging==25.0
phonenumbers==9.0.18
pillow==12.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
redis==7.0.1
referencing==0.37.0
request-id==1.0.1
requests==2.32.5
rpds-py==0.29.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
swapper==1.4.0
twilio==9.8.6
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.41.0
vine==5.1.0
wcwidth==0.2.14
WebOb==1.8.9
yarl==1.22.0
